gates and policies

+++ app/Http/Controllers/ContactController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\ContactResource;
 use App\Models\Contact;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -12,19 +13,31 @@ class ContactController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $contacts = Contact::latest()->get();
+        if ($request->user()->cannot('viewAny', Contact::class)) {
+            abort(403);
+        }
+
+        $contacts = Contact::query();
+        if($request->user()->role == 'user')
+        {
+            $contacts = $contacts->where('user_id', $request->user()->id);
+        }
+        $contacts = $contacts->latest()->get();
         return Inertia::render('contacts/index', [
-            'contacts' => $contacts,
+            'contacts' => ContactResource::collection($contacts),
         ]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        if ($request->user()->cannot('create', Contact::class)) {
+            abort(403);
+        }
         return Inertia::render('contacts/create');
     }
 
@@ -33,6 +46,9 @@ class ContactController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->user()->cannot('create', Contact::class)) {
+            abort(403);
+        }
         $validatedData = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -57,8 +73,11 @@ class ContactController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Contact $contact)
+    public function edit(Request $request, Contact $contact)
     {
+        if ($request->user()->cannot('update', $contact)) {
+            abort(403);
+        }
         return Inertia::render('contacts/edit', [
             'contact' => $contact,
         ]);
@@ -69,6 +88,10 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
+        if ($request->user()->cannot('update', $contact)) {
+            abort(403);
+        }
+
         $validatedData = $request->validate([
             'name' => 'required',
             'email' => 'nullable|email',
@@ -84,8 +107,12 @@ class ContactController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Contact $contact)
+    public function destroy(Request $request, Contact $contact)
     {
+        if ($request->user()->cannot('delete', $contact)) {
+            abort(403);
+        }
+
         $contact->delete();
 
         return redirect()->route('contacts.index')->with(['success', 'Contact deleted successfully']);
